Rename ghub to issues
Rename to  issues
