+++ issues/info.json
@@ -2,8 +2,8 @@
   "author": [
     "TurnrDev"
   ],
-  "description": "Create, edit and close your GitHub issues",
-  "short": "Create, edit and close your GitHub issues",
+  "description": "Create and view your GitHub issues",
+  "short": "Create and view your GitHub issues",
   "min_bot_version": "3.4.0",
   "requirements": [
     "discord-ext-alternatives>=2020.6.16"
